adding ejb

+++ src/main/java/com/park/parkinglot2/entity/Car.java
@@ -30,7 +30,7 @@ public class Car implements Serializable {
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
 
     private String licensePlate;
 
@@ -66,14 +66,16 @@ public class Car implements Serializable {
     public void setUser(User user) {
         this.user = user;
     }
-    
-    public Long getId() {
+
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
+    
+    
 
     @Override
     public int hashCode() {
